Added perfomance tests
In these tests, after hosting a server, clients can connect and issue
a certain number of requests.
For testing the server, the client creates a certain number of threads
that send get/put requests to the server. They measure the response time
and count the number of requests to which they successfully received a
response.
Parameters for the test can be specified in tests/test_common.h